# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Todo List",
        "description": "Dokumentasi api yang tersedia",
        "contact": {
            "email": "iqbal.ramadhani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/show-detail-checklist-item": {
            "get": {
                "tags": [
                    "Check List Item"
                ],
                "summary": "Show All Checklist Item",
                "description": "Show All Checklist Item",
                "operationId": "30c5aaa14b4232cd8755bac4a69a7a38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID checklist",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Error validate access token or form data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-checklist-item": {
            "post": {
                "tags": [
                    "Check List Item"
                ],
                "summary": "Add Checklist Item",
                "description": "",
                "operationId": "13ab71a8143ae737c3301b7b31f86bf3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID Check List",
                                        "type": "number"
                                    },
                                    "item": {
                                        "description": "Nama Item",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Data Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Auth Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/remove-checklist-item": {
            "delete": {
                "tags": [
                    "Check List Item"
                ],
                "summary": "Remove Checklist",
                "description": "",
                "operationId": "12779e7e0709980149928cbea893c55d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID Checklist Item",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Data Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Auth Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-checklist-item": {
            "put": {
                "tags": [
                    "Check List Item"
                ],
                "summary": "Upde Status Checklist Item",
                "description": "",
                "operationId": "034f142170056a1bb94a22cfc2aa4b62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID checklist item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Data Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Auth Validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}